allow switch on/off openql/audio

+++ ui/run.go
@@ -25,38 +25,52 @@ func init() {
 }
 
 func Run(paths []string) {
-	// initialize audio
-	portaudio.Initialize()
-	defer portaudio.Terminate()
+	var (
+		glDisabled = false
+		audioDisabled = true
+		window *glfw.Window
+		audio *Audio
+	)
 
-	audio := NewAudio()
-	if err := audio.Start(); err != nil {
-		log.Fatalln(err)
-	}
-	defer audio.Stop()
+	if !audioDisabled {
+		// initialize audio
+		portaudio.Initialize()
+		defer portaudio.Terminate()
 
-	// initialize glfw
-	if err := glfw.Init(); err != nil {
-		log.Fatalln(err)
+		audio := NewAudio()
+		if err := audio.Start(); err != nil {
+			log.Fatalln(err)
+		}
+		defer audio.Stop()
 	}
-	defer glfw.Terminate()
 
-	// create window
-	glfw.WindowHint(glfw.ContextVersionMajor, 2)
-	glfw.WindowHint(glfw.ContextVersionMinor, 1)
-	window, err := glfw.CreateWindow(width*scale, height*scale, title, nil, nil)
-	if err != nil {
-		log.Fatalln(err)
-	}
-	window.MakeContextCurrent()
 
-	// initialize gl
-	if err := gl.Init(); err != nil {
-		log.Fatalln(err)
+	if !glDisabled {
+		// initialize glfw
+		if err := glfw.Init(); err != nil {
+			log.Fatalln(err)
+		}
+		defer glfw.Terminate()
+
+		// create window
+		glfw.WindowHint(glfw.ContextVersionMajor, 2)
+		glfw.WindowHint(glfw.ContextVersionMinor, 1)
+		var err error
+		window, err = glfw.CreateWindow(width*scale, height*scale, title, nil, nil)
+		if err != nil {
+			log.Fatalln(err)
+		}
+		window.MakeContextCurrent()
+
+		// initialize gl
+		if err := gl.Init(); err != nil {
+			log.Fatalln(err)
+		}
+		gl.Enable(gl.TEXTURE_2D)
 	}
-	gl.Enable(gl.TEXTURE_2D)
+
 
 	// run director
-	director := NewDirector(window, audio)
+	director := NewDirector(window, audio, glDisabled, audioDisabled)
 	director.Start(paths)
 }
